+++ SimpleManager.h
@@ -12,9 +12,9 @@ public:
 	Manager();
 	Manager(string name);
 
-	double calculateAvgYearsOfWork(Group group);
+	int calculateAvgYearsOfWork(Group group);
 	int calculateHoursEveryWeek(Group group);
-	double calculateAvgAge(Group group);
+	int calculateAvgAge(Group group);
 	int getCountOfMale(Group group);
 	int getCountOfFemale(Group group);
 	;
